# BUG: fix regex to avoid DoS