--- a/nltk/corpus/reader/xmldocs.py
+++ b/nltk/corpus/reader/xmldocs.py
@@ -331,7 +331,7 @@ class XMLCorpusView(StreamBackedCorpusView):
                     print("%25s %s" % ("/".join(context)[-20:], piece.group()))
 
                 if piece.group("START_TAG"):
-                    name = self._XML_TAG_NAME.match(piece.group()).group(1)
+                    name = self._XML_TAG_NAME.match(piece.group()).group(2)
                     # Keep context up-to-date.
                     context.append(name)
                     # Is this one of the elts we're looking for?
@@ -341,7 +341,7 @@ class XMLCorpusView(StreamBackedCorpusView):
                             elt_depth = len(context)
 
                 elif piece.group("END_TAG"):
-                    name = self._XML_TAG_NAME.match(piece.group()).group(1)
+                    name = self._XML_TAG_NAME.match(piece.group()).group(2)
                     # sanity checks:
                     if not context:
                         raise ValueError("Unmatched tag </%s>" % name)
@@ -359,7 +359,7 @@ class XMLCorpusView(StreamBackedCorpusView):
                     context.pop()
 
                 elif piece.group("EMPTY_ELT_TAG"):
-                    name = self._XML_TAG_NAME.match(piece.group()).group(1)
+                    name = self._XML_TAG_NAME.match(piece.group()).group(2)
                     if elt_start is None:
                         if re.match(tagspec, "/".join(context) + "/" + name):
                             elts.append((piece.group(), "/".join(context) + "/" + name))