--- a/nltk/corpus/reader/xmldocs.py
+++ b/nltk/corpus/reader/xmldocs.py
@@ -222,8 +222,7 @@ class XMLCorpusView(StreamBackedCorpusView):
 
     #: A regular expression used to extract the tag name from a start tag,
     #: end tag, or empty-elt tag string.
-    # _XML_TAG_NAME = re.compile("<\s*/?\s*([^\s>]+)")
-    _XML_TAG_NAME = re.compile("<\s*(/\s*)?([^\s>]+)")
+    _XML_TAG_NAME = re.compile("<\s*/?\s*([^\s>]+)")
 
     #: A regular expression used to find all start-tags, end-tags, and
     #: emtpy-elt tags in an XML file.  This regexp is more lenient than