The regex which can match over long parts of strings but then be rejected by nonmatching at the end can be used for
denial-of-service. This addresses #2596