@yetingli 2020-09-13
Well done, I can't agree with you more!  Thanks @CubieDev :)
I will revert commit and thanks again.