@CubieDev 2020-09-24
These tests failed due to inconsistencies with the Python 3.5 UnigramTagger. I'm investigating it now.